+++ teste/Podfile
@@ -2,7 +2,7 @@
 platform :ios, '9.0'
 # ignore all warnings from all pods
 inhibit_all_warnings!
-target 'teste' do
+target 'lodjaApp' do
   # Comment the next line if you're not using Swift and don't want to use dynamic frameworks
   use_frameworks!
