+++ teste/lodjaApp.xcworkspace/contents.xcworkspacedata
@@ -2,7 +2,7 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:/Users/rmarangoni/Documents/Diversos/challenge-ios/teste/lodjaApp.xcodeproj">
+      location = "group:lodjaApp.xcodeproj">
    </FileRef>
    <FileRef
       location = "group:Pods/Pods.xcodeproj">
