+++ teste/teste/Extensions.swift
@@ -1,6 +1,6 @@
 //
 //  Extensions.swift
-//  teste
+//  lodjaApp
 //
 //  Created by Rodrigo Marangoni on 30/01/18.
 //  Copyright © 2018 romarangoni. All rights reserved.
@@ -14,4 +14,11 @@ extension UILabel {
         attributeString.addAttribute(NSAttributedStringKey.strikethroughStyle, value: 2, range: NSMakeRange(0, attributeString.length))
         self.attributedText = attributeString;
     }
+}
+
+extension UIColor {
+    static func purpleB2W() -> UIColor {
+        return UIColor(red: 96.0/256, green: 44.0/256, blue: 127.0/256, alpha: 1)
+    }
+    
 }
