+++ teste/teste/Model/Category.swift
@@ -1,34 +1,32 @@
 //
-//  Banner.swift
-//  teste
+//  Category.swift
+//  lodjaApp
 //
 //  Created by Rodrigo Marangoni on 30/01/18.
 //  Copyright © 2018 romarangoni. All rights reserved.
 //
 
 import Foundation
-
-import Foundation
 import ObjectMapper
 
-class Banner: Mappable {
-    var bannerId: Int?
+class Category: Mappable {
+    var descricao: String?
+    var catId: Int?
     var urlImage: String?
-    var linkUrl: String?
     
     required init?(map: Map) {
         
     }
     func mapping(map: Map) {
-        bannerId <- map["id"]
+        descricao <- map["descricao"]
+        catId <- map["id"]
         urlImage <- map["urlImagem"]
-        linkUrl <- map["linkUrl"]
     }
     
 }
 
-class BannerData: Mappable{
-    var data:[Banner]?
+class CategoryData: Mappable{
+    var data:[Category]?
     required init?(map: Map) {
         mapping(map: map)
     }
