+++ teste/teste/Model/Product.swift
@@ -1,6 +1,6 @@
 //
-//  Category.swift
-//  teste
+//  Product.swift
+//  lodjaApp
 //
 //  Created by Rodrigo Marangoni on 30/01/18.
 //  Copyright © 2018 romarangoni. All rights reserved.
@@ -9,24 +9,31 @@
 import Foundation
 import ObjectMapper
 
-class Category: Mappable {
+class Product: Mappable {
+    var category: Category?
     var descricao: String?
-    var catId: Int?
+    var prodId: Int?
     var urlImage: String?
+    var nome: String?
+    var precoDe: Float?
+    var precoPor: Float?
     
     required init?(map: Map) {
         
     }
     func mapping(map: Map) {
+        category <- map["categoria"]
         descricao <- map["descricao"]
-        catId <- map["id"]
+        prodId <- map["id"]
         urlImage <- map["urlImagem"]
+        nome <- map["nome"]
+        precoDe <- map["precoDe"]
+        precoPor <- map["precoPor"]
     }
-    
 }
 
-class CategoryData: Mappable{
-    var data:[Category]?
+class ProductData: Mappable{
+    var data:[Product]?
     required init?(map: Map) {
         mapping(map: map)
     }
