+++ teste/teste/ViewControllers/HomeViewController.swift
@@ -1,6 +1,6 @@
 //
 //  HomeViewController.swift
-//  teste
+//  lodjaApp
 //
 //  Created by Rodrigo Marangoni on 26/01/18.
 //  Copyright © 2018 romarangoni. All rights reserved.
