+++ teste/teste/Views/CategoryView.swift
@@ -1,6 +1,6 @@
 //
 //  CategoryView.swift
-//  teste
+//  lodjaApp
 //
 //  Created by Rodrigo Marangoni on 29/01/18.
 //  Copyright © 2018 romarangoni. All rights reserved.
