+++ teste/teste/Views/CategoryView.xib
@@ -12,7 +12,7 @@
     <objects>
         <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
         <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view contentMode="scaleToFill" id="iN0-l3-epB" customClass="CategoryView" customModule="teste" customModuleProvider="target">
+        <view contentMode="scaleToFill" id="iN0-l3-epB" customClass="CategoryView" customModule="lodjaApp" customModuleProvider="target">
             <rect key="frame" x="0.0" y="0.0" width="120" height="120"/>
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
             <subviews>
@@ -25,10 +25,10 @@
                     </constraints>
                 </imageView>
                 <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Descrição" textAlignment="center" lineBreakMode="wordWrap" numberOfLines="2" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="Wk8-oS-ynE">
-                    <rect key="frame" x="4" y="82" width="112" height="30"/>
-                    <fontDescription key="fontDescription" type="system" pointSize="11"/>
+                    <rect key="frame" x="2" y="82" width="116" height="30"/>
+                    <fontDescription key="fontDescription" type="system" pointSize="12"/>
                     <nil key="textColor"/>
-                    <nil key="highlightedColor"/>
+                    <color key="highlightedColor" red="0.20000000000000001" green="0.20000000000000001" blue="0.20000000000000001" alpha="1" colorSpace="calibratedRGB"/>
                 </label>
                 <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="zig-yt-bd5">
                     <rect key="frame" x="0.0" y="0.0" width="120" height="120"/>
@@ -37,10 +37,10 @@
             <constraints>
                 <constraint firstItem="zig-yt-bd5" firstAttribute="top" secondItem="vUN-kp-3ea" secondAttribute="top" id="3TE-Jz-gtX"/>
                 <constraint firstAttribute="bottom" secondItem="zig-yt-bd5" secondAttribute="bottom" id="MEX-Fx-vHf"/>
-                <constraint firstItem="vUN-kp-3ea" firstAttribute="trailing" secondItem="Wk8-oS-ynE" secondAttribute="trailing" constant="4" id="Mei-Tu-RBp"/>
+                <constraint firstItem="vUN-kp-3ea" firstAttribute="trailing" secondItem="Wk8-oS-ynE" secondAttribute="trailing" constant="2" id="Mei-Tu-RBp"/>
                 <constraint firstItem="Wk8-oS-ynE" firstAttribute="top" secondItem="L9v-QI-oIB" secondAttribute="bottom" constant="4" id="NJc-mi-V1H"/>
                 <constraint firstItem="vUN-kp-3ea" firstAttribute="trailing" secondItem="zig-yt-bd5" secondAttribute="trailing" id="Zea-du-4VN"/>
-                <constraint firstItem="Wk8-oS-ynE" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" constant="4" id="bBQ-W7-eTF"/>
+                <constraint firstItem="Wk8-oS-ynE" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" constant="2" id="bBQ-W7-eTF"/>
                 <constraint firstItem="vUN-kp-3ea" firstAttribute="bottom" secondItem="Wk8-oS-ynE" secondAttribute="bottom" constant="8" id="gbf-Ug-BL8"/>
                 <constraint firstItem="zig-yt-bd5" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" id="meg-X7-xHO"/>
                 <constraint firstItem="Wk8-oS-ynE" firstAttribute="centerX" secondItem="iN0-l3-epB" secondAttribute="centerX" id="sa1-tl-bhR"/>
